fix bug

+++ generate-voice-local.py
@@ -255,7 +255,7 @@ def get_tts_wav(ref_wav_path, prompt_text, prompt_language, text, text_language,
     prompt_language = prompt_language
     text_language = text_language
     phones1, word2ph1, norm_text1 = get_cleaned_text_final(prompt_text, prompt_language)
-    bert1=get_bert_final(phones1, word2ph1, norm_text1,prompt_language,g_para.device).to(g_para.dtype)
+    bert1=get_bert_final(phones1, word2ph1, norm_text1,prompt_language,g_para.device).to(torch.float16)
     texts = text.split("\n")
     audio_opt = []
